Implemented temporary way of downloading single files and displaying the contents of a FileEntity
This code is functional, although the base system itself would enjoy some rework. Do not hesitate on using this code though, as all future changes to this system should be compatible with the current way of dealing with files.

P.S. I also made the code a bit more maintainable.